+++ Renderer/Renderer/RayCasting/Geometry/RayTraceScene.h
@@ -47,7 +47,9 @@ public:
 	void			   AddObject(std::shared_ptr<I_RayGeometryObject>&& object);
 	void			   AddLight(std::shared_ptr<RayTracing::C_AreaLight>&& light);
 	void			   AddLight(std::shared_ptr<RayTracing::C_PointLight>&& light);
-	void			   AddMesh(const Core::ResourceHandle<C_TrimeshModel> trimesh, const MeshData::Material& material);
+	void			   AddMesh(const Core::ResourceHandle<C_TrimeshModel>& trimesh,
+							   const MeshData::Material&                   material,
+							   const glm::mat4&                            transform = glm::mat4(1.f));
 
 	void ForEachLight(const std::function<void(const std::reference_wrapper<const RayTracing::I_RayLight>& light)>& fnc) const;
 
@@ -58,6 +60,8 @@ public:
 	[[nodiscard]] bool IsLoaded() const;
 	void			   BuildScene(); //< Call after fully loaded
 	void			   ClearScene();
+
+	std::unique_ptr<I_MaterialInterface>& AddMaterial(const MeshData::Material& material);
 
 private:
 	std::vector<std::shared_ptr<I_RayGeometryObject>>	   m_Objects;
@@ -73,8 +77,6 @@ private:
 	Core::LoadingQuery m_LoadingMeshes;
 	Core::LoadingQuery m_LoadingBVH;
 	Core::LoadingQuery m_LoadingTextures;
-
-	std::unique_ptr<I_MaterialInterface>& AddMaterial(const MeshData::Material& material);
 };
 
 } // namespace GLEngine::Renderer