+++ VulkanRenderer/VulkanRenderer/VkWindow.cpp
@@ -765,7 +765,8 @@ void C_VkWindow::SetupGUI()
 	{
 		m_ImageEditorGUID = NextGUID();
 
-		auto* imageEditorWindow = new Editor::C_ImageEditor(m_ImageEditorGUID, guiMGR);
+		auto* imageEditorWindow = new Editor::C_ImageEditor(m_ImageEditorGUID, guiMGR,
+			[this](Core::I_Event& e) { OnEvent(e); });
 
 		guiMGR.AddCustomWindow(imageEditorWindow);
 		imageEditorWindow->SetVisible();